Feature: About Us page link from Home page & split Home from Root-config (#387)
* fix: remove custom sw

* feat: add link to aboutus page from home

* fix: eslint
