+++ root/src/constants/locales/en/translation.json
@@ -10,6 +10,7 @@
     "feeds": "Feeds"
   },
   "aboutus": {
+    "title": "About Us",
     "our mission": "Our Mission",
     "join us": "Join Us",
     "meet our team": "Meet Our Team",
