+++ root/src/constants/locales/jp/translation.json
@@ -10,6 +10,7 @@
     "feeds": "記事"
   },
   "aboutus": {
+    "title": "チーム概要",
     "our mission": "理念",
     "join us": "募集",
     "meet our team": "メンバー",
